+++ src/MEII/Utility/VirtualInput.cpp
@@ -7,7 +7,7 @@ namespace meii {
 
 	VirtualInput::VirtualInput(const std::string& name, std::vector<mel::uint32> channel_numbers) :
 		seed_(std::chrono::system_clock::now().time_since_epoch().count()),
-		generator_(seed_),
+		generator_((unsigned int)seed_),
 		distribution_(0.0,1.0)
 	{
 		set_name(name);
